Cambio de dato a tipo float64 para la columna edad

+++ titanic.py
@@ -2,6 +2,6 @@ import pandas as pd
 import numpy as np
 import matplotlib.pyplot as plt
 
-data = pd.read_csv('datasets/titanic3.csv', dtype={'age':np.int32})
+data = pd.read_csv('datasets/titanic3.csv', dtype={'age':np.float64})
 
 print(data.head())
